Moved remote servers loading to separated action
