+++ test/servers/reducers/servers.test.js
@@ -2,10 +2,8 @@ import { values } from 'ramda';
 import reducer, {
   createServer,
   deleteServer,
-  listServers,
   createServers,
   editServer,
-  LIST_SERVERS,
   EDIT_SERVER,
   DELETE_SERVER,
   CREATE_SERVERS,
@@ -16,21 +14,10 @@ describe('serverReducer', () => {
     abc123: { id: 'abc123' },
     def456: { id: 'def456' },
   };
-  const expectedFetchServersResult = { type: LIST_SERVERS, list };
-  const ServersServiceMock = {
-    listServers: jest.fn(() => list),
-    createServer: jest.fn(),
-    editServer: jest.fn(),
-    deleteServer: jest.fn(),
-    createServers: jest.fn(),
-  };
 
   afterEach(jest.clearAllMocks);
 
   describe('reducer', () => {
-    it('returns servers when action is LIST_SERVERS', () =>
-      expect(reducer({}, { type: LIST_SERVERS, list })).toEqual(list));
-
     it('returns edited server when action is EDIT_SERVER', () =>
       expect(reducer(
         list,
@@ -59,75 +46,6 @@ describe('serverReducer', () => {
   });
 
   describe('action creators', () => {
-    xdescribe('listServers', () => {
-      const axios = { get: jest.fn() };
-      const dispatch = jest.fn();
-      const NoListServersServiceMock = { ...ServersServiceMock, listServers: jest.fn(() => ({})) };
-
-      it('fetches servers from local storage when found', async () => {
-        await listServers(ServersServiceMock, axios)()(dispatch);
-
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenNthCalledWith(1, expectedFetchServersResult);
-        expect(ServersServiceMock.listServers).toHaveBeenCalledTimes(1);
-        expect(ServersServiceMock.createServer).not.toHaveBeenCalled();
-        expect(ServersServiceMock.editServer).not.toHaveBeenCalled();
-        expect(ServersServiceMock.deleteServer).not.toHaveBeenCalled();
-        expect(ServersServiceMock.createServers).not.toHaveBeenCalled();
-        expect(axios.get).not.toHaveBeenCalled();
-      });
-
-      it.each([
-        [
-          Promise.resolve({
-            data: [
-              {
-                id: '111',
-                name: 'acel.me from servers.json',
-                url: 'https://acel.me',
-                apiKey: '07fb8a96-8059-4094-a24c-80a7d5e7e9b0',
-              },
-              {
-                id: '222',
-                name: 'Local from servers.json',
-                url: 'http://localhost:8000',
-                apiKey: '7a531c75-134e-4d5c-86e0-a71b7167b57a',
-              },
-            ],
-          }),
-          {
-            111: {
-              id: '111',
-              name: 'acel.me from servers.json',
-              url: 'https://acel.me',
-              apiKey: '07fb8a96-8059-4094-a24c-80a7d5e7e9b0',
-            },
-            222: {
-              id: '222',
-              name: 'Local from servers.json',
-              url: 'http://localhost:8000',
-              apiKey: '7a531c75-134e-4d5c-86e0-a71b7167b57a',
-            },
-          },
-        ],
-        [ Promise.resolve('<html></html>'), {}],
-        [ Promise.reject({}), {}],
-      ])('tries to fetch servers from remote when not found locally', async (mockedValue, expectedList) => {
-        axios.get.mockReturnValue(mockedValue);
-
-        await listServers(NoListServersServiceMock, axios)()(dispatch);
-
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenNthCalledWith(1, { type: LIST_SERVERS, list: expectedList });
-        expect(NoListServersServiceMock.listServers).toHaveBeenCalledTimes(1);
-        expect(NoListServersServiceMock.createServer).not.toHaveBeenCalled();
-        expect(NoListServersServiceMock.editServer).not.toHaveBeenCalled();
-        expect(NoListServersServiceMock.deleteServer).not.toHaveBeenCalled();
-        expect(NoListServersServiceMock.createServers).toHaveBeenCalledTimes(1);
-        expect(axios.get).toHaveBeenCalledTimes(1);
-      });
-    });
-
     describe('createServer', () => {
       it('returns expected action', () => {
         const serverToCreate = { id: 'abc123' };
